requirements.txt
fastapi
sqlalchemy
uvicorn
pydantic[email]
psycopg2
bcrypt
passlib
pyjwt
jose